pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fingerprint-squared"
version = "0.1.0"
description = "Fingerprint²: A Comprehensive Ethical AI Assessment Framework for Vision-Language Models"
readme = "README.md"
license = {text = "Apache-2.0"}
authors = [
    {name = "Fingerprint² Research Team"}
]
keywords = [
    "fairness",
    "bias",
    "vlm",
    "vision-language-models",
    "ethical-ai",
    "responsible-ai",
    "benchmark",
    "evaluation"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
requires-python = ">=3.9"
dependencies = [
    # Core ML
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "transformers>=4.45.0",  # Updated for LLaMA 3.2 Vision
    "datasets>=2.14.0",
    "pillow>=9.0.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "scipy>=1.10.0",
    "scikit-learn>=1.3.0",

    # Visualization
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "plotly>=5.15.0",

    # Utilities
    "tqdm>=4.65.0",
    "pyyaml>=6.0",
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    "wandb>=0.15.0",

    # API clients
    "openai>=1.0.0",
    "anthropic>=0.18.0",
    "google-generativeai>=0.3.0",
    "replicate>=0.20.0",
    "httpx>=0.24.0",  # For OpenRouter
    "aiohttp>=3.8.0",
    "tenacity>=8.2.0",

    # CLI
    "rich>=13.0.0",
    "typer>=0.9.0",

    # Data/IO
    "jsonlines>=3.1.0",
    "huggingface-hub>=0.19.0",
    "reportlab>=4.0.0",

    # Qwen VL utils (optional but recommended)
    "qwen-vl-utils>=0.0.2;python_version>='3.9'",

    # Async support
    "tqdm[asyncio]>=4.65.0",

    # API server
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
    "ruff>=0.0.290",
]
notebooks = [
    "jupyter>=1.0.0",
    "ipywidgets>=8.1.0",
    "nbformat>=5.9.0",
]
# Flash attention for faster inference (requires CUDA)
flash = [
    "flash-attn>=2.5.0",
]
# vLLM for high-throughput batch inference
vllm = [
    "vllm>=0.4.0",
]
# Face detection for preprocessing
face_detection = [
    "opencv-python>=4.8.0",
    "mtcnn>=0.1.1",
    "mediapipe>=0.10.0",
]
all = [
    "fingerprint-squared[dev,notebooks,flash,face_detection]",
]

[project.scripts]
fingerprint = "fingerprint_squared.cli.main:main"
fp2 = "fingerprint_squared.cli.main:main"
fingerprint-squared = "fingerprint_squared.cli.main:main"
fp2-server = "fingerprint_squared.api.server:main"

[project.urls]
Homepage = "https://github.com/fingerprint-squared/fingerprint-squared"
Documentation = "https://fingerprint-squared.readthedocs.io"
Repository = "https://github.com/fingerprint-squared/fingerprint-squared"

[tool.setuptools.packages.find]
where = ["."]
include = ["fingerprint_squared*"]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311"]

[tool.isort]
profile = "black"
line_length = 100

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
